auth_service/requirements.txt
Flask==2.3.3
pymongo==4.5.0
PyJWT==2.8.0 